Added long standing process to test rejoining.
Converted structure to a handler class that holds state to make the connections easier to store.

Might need to pass a logger to each process so they can use it separately and hold different streams with different locks and different buffers, etc. It might also be beneficial to store sessions by process rather than by client.